+++ Src/Idioms/Idioms.csproj
@@ -2,7 +2,7 @@
   <Import Project="..\..\Common.props"/>
 
   <PropertyGroup>
-    <TargetFrameworks>net6.0</TargetFrameworks>
+    <TargetFrameworks>net462;netstandard2.0;net5.0;net6.0</TargetFrameworks>
     <AssemblyTitle>AutoFixture.Idioms</AssemblyTitle>
     <AssemblyName>AutoFixture.Idioms</AssemblyName>
     <RootNamespace>AutoFixture.Idioms</RootNamespace>
