+++ Src/IdiomsUnitTest/IdiomsUnitTest.csproj
@@ -4,7 +4,7 @@
   <Import Project="..\..\Common.Test.xUnit.props" />
 
   <PropertyGroup>
-    <TargetFrameworks>net6.0</TargetFrameworks>
+    <TargetFrameworks>net462;net48;net5.0;net6.0</TargetFrameworks>
     <AssemblyTitle>IdiomsUnitTest</AssemblyTitle>
     <AssemblyName>AutoFixture.IdiomsUnitTest</AssemblyName>
     <RootNamespace>AutoFixture.IdiomsUnitTest</RootNamespace>
